refactor: rename files in email module
