+++ src/modules/Email/Adapter/Interfaces/IAdapter.ts
