--- src/modules/Email/Core/ICore.ts
@@ -1,3 +1,0 @@
-import { IAdapter } from "../Adapter/IAdapter";
-
-export type ICore = IAdapter;
