+++ src/modules/Email/Core/IEmailCore.ts
@@ -1,0 +1,3 @@
+import { IAdapter } from "../Adapter/Interfaces/IAdapter";
+
+export type IEmailCore = IAdapter;
